// app/models/settings.ts
export class Settings {
	// online
	backgroundColorOnline: string;
	foregroundColorOnline: string;
	blinkOnline: boolean;
	
	// away
	backgroundColorAway: string;
	foregroundColorAway: string;
	blinkAway: boolean;
	
	// busy
	backgroundColorBusy: string;
	foregroundColorBusy: string;
	blinkBusy: boolean;
	
	// dnd
	backgroundColorDnd: string;
	foregroundColorDnd: string;
	blinkDnd: boolean;
	
	// undefined
	backgroundColorUndefined: string;
	foregroundColorUndefined: string;
	blinkUndefined: boolean;
}